moving recon utils into utils.py
