+++ utils.py
@@ -1,10 +1,13 @@
 import hashlib
 import numpy as np
 import os
+import glob
 import pydicom
 import random
 import uuid
 import scipy
+import matplotlib.pyplot as plt
+from pathlib import Path
 from scipy.ndimage.morphology import binary_erosion
 from scipy.ndimage.morphology import binary_dilation
 from scipy import spatial
@@ -440,7 +443,6 @@ def all_points_merged(poly, merged):
     return total_pts.shape[0] <= merged.shape[0]
 
 
-# TODO: Make two poly coords functions for MIM or Pinnacle
 def poly_to_coords_2D(poly, ctcoord, flatten=True, invert=False):
     """
     Function
@@ -601,10 +603,86 @@ def img_dims(dicom_list):
         else:
             loc_list += list(loc0 + np.arange(1, diff + 1) * thickness)
         # Resorts the list with the new points
-        int_list, loc_list = map(np.array, zip(
-            *sorted(zip(int_list, loc_list))))
+        int_list, loc_list = map(np.array, zip(*sorted(zip(int_list, loc_list))))
 
-    return thickness, n_slices, loc_list.min(), loc_list.max(), flip
+    return thickness, int(n_slices), loc_list.min(), loc_list.max(), flip
+
+
+def nearest(array, value):
+    """
+    Function
+    ----------
+    Finds the nearest index of array to a given value
+
+    Parameters
+    ----------
+    array : np.array
+    value : float
+
+    Returns
+    ----------
+    index : int
+    """
+    return np.abs(np.asarray(array) - value).argmin()
+
+
+def key_list_creator(key_list, *args):
+    """
+    Function
+    ----------
+    Smaller wrapper to create a key ordering list for the rtstruct func
+
+    Parameters
+    ----------
+    key_list : list
+        A list of the desired index order of structures, if present
+    *args : list of arguments
+        Arguements to be passed to key_list.index(*)
+
+    Returns
+    ----------
+    Returns a function which yeilds the proper index of each structure
+    """
+    if type(key_list) is dict:
+        key_list = list(key_list.keys())
+
+    def new_key(*args):
+        return key_list.index(*args)
+    return new_key
+
+
+def slice_thickness(dcm0, dcm1):
+    """
+    Function
+    ----------
+    Computes the slice thickness for a DICOM set
+
+    Parameters
+    ----------
+    dcm0, dcm1 : str or pydicom.dataset.FileDataset
+        Either a string to the dicom path or a pydicom dataset
+
+    Returns
+    ----------
+    slice_thickness : float
+        A float representing the robustly calculated slice thickness
+
+    Notes
+    ----------
+        Calculates based on slice location and instance number.
+        Does not trust SliceThickness DICOM Header
+    """
+    if type(dcm0) != pydicom.dataset.FileDataset:
+        dcm0 = pydicom.dcmread(dcm0)
+    if type(dcm1) != pydicom.dataset.FileDataset:
+        dcm1 = pydicom.dcmread(dcm1)
+
+    loc0 = dcm0.SliceLocation
+    loc1 = dcm1.SliceLocation
+    inst0 = dcm0.InstanceNumber
+    inst1 = dcm1.InstanceNumber
+
+    return abs((loc1-loc0) / (inst1-inst0))
 
 
 def generate_instance_uid():
@@ -645,3 +723,341 @@ def generate_instance_uid():
     # Format all the SOP Instance UID stops properly
     terms = hv[:6], hv[6], hv[7:15], hv[15:26], hv[26:35], hv[35:38], hv[38:41]
     return '2.16.840.1.' + '.'.join(terms)
+
+
+def d_max_coords(patient_path, dose_volume, printing=True):
+    """
+    Function
+    ----------
+    Used to determine the coordinates of D_max, as a dose reconstruction
+        sanity check. Can be compared against the TPS or MIM
+
+    Parameters
+    ----------
+    patient_path : str
+        A path directing towards a patient database in the following format:
+            MRN/[CT,PET,RTSTRUCT,RTDOSE]/*.dcm
+    dose_volume : array
+        An array of the dose volume, as returned form the dose
+            reconstruction function
+    printing : bool (Default = True)
+        Prints a formatting of the results, along with the voxel size
+
+    Returns
+    ----------
+    (volume_max, ct_max_mm, dose_max_mm, ct_voxel_size) : (float, float, float, float)
+        0.The dose maximum coordinates (x, y, z) in voxels of the volume
+        1.The dose maximum coordinates (x, y, z) in mm relative to isocenter
+            in the CT coordinate system
+        2.The dose maximum coordinates (x, y, z) in mm relative to isocenter
+            in the RTDOSE coordinate system
+        3.The CT voxel size for (x, y, z) in mm
+
+    Notes
+    ----------
+    Due to the RTDOSE computation, it is likely in a slighly different
+        coordinate system than the CT coordinates. But, the slice difference
+        should be < 1/2 * voxel size of the CT coordinates
+    """
+    if patient_path[0] == '~':
+        patient_path = os.path.expanduser('~') + patient_path[1:]
+
+    if patient_path[-1] != '/':
+        patient_path += '/'
+
+    dose_files = glob.glob(patient_path + 'RTDOSE/*.dcm')
+    ct_files = glob.glob(patient_path + 'CT/*.dcm')
+    ct_files.sort()
+
+    dose_dcm = pydicom.dcmread(dose_files[0])
+    for ct_file in ct_files:
+        ct_dcm = pydicom.dcmread(ct_file, stop_before_pixels=True)
+        if ct_dcm.InstanceNumber == 1:
+            z_0 = float(ct_dcm.ImagePositionPatient[-1])
+        if ct_dcm.InstanceNumber == len(ct_files):
+            z_1 = float(ct_dcm.ImagePositionPatient[-1])
+
+    img_dims = (ct_dcm.Rows, ct_dcm.Columns, len(ct_files))
+    img_origin = np.array([*ct_dcm.ImagePositionPatient[:2], min(z_0, z_1)])
+    ix, iy, iz = (*ct_dcm.PixelSpacing, ct_dcm.SliceThickness)
+    ct_voxel_size = np.array([ix, iy, iz])
+
+    dose_dims = np.rollaxis(dose_dcm.pixel_array, 0, 3).shape
+    dose_iso = np.array(dose_dcm.ImagePositionPatient)
+    dx, dy, dz = (*dose_dcm.PixelSpacing, dose_dcm.SliceThickness)
+
+    d_grid_x = dose_iso[1]+dx * np.arange(dose_dims[0])
+    d_grid_y = dose_iso[0]+dy * np.arange(dose_dims[1])
+    d_grid_z = dose_iso[2]+dz * np.arange(dose_dims[2])
+
+    i_grid_x = img_origin[1]+ix * np.arange(img_dims[0])
+    i_grid_y = img_origin[0]+iy * np.arange(img_dims[1])
+    i_grid_z = img_origin[2]+iz * np.arange(img_dims[2])
+
+    if z_0 < z_1:
+        i_grid_z = i_grid_z[::-1]
+
+    volume_max = np.unravel_index(dose_volume.argmax(),
+                                  dose_volume.shape)
+
+    ct_max_mm = np.array([i_grid_x[volume_max[0]],
+                          i_grid_y[volume_max[1]],
+                          i_grid_z[volume_max[2]]])
+
+    dose_coords = np.unravel_index(dose_dcm.pixel_array.argmax(),
+                                   dose_dcm.pixel_array.shape)
+
+    dose_max_mm = np.array([d_grid_x[dose_coords[1]],
+                            d_grid_y[dose_coords[2]],
+                            d_grid_z[dose_coords[0]]])
+
+    if printing:
+        with np.printoptions(formatter={'float': '{:>6.2f} '.format,
+                                        'int': '{:>3d}    '.format}):
+            print(f'dose_volume max (voxels): {np.array(volume_max)}')
+            print(f'D_max CT coordinate (mm): {ct_max_mm}')
+            print(f'D_max RT coordinate (mm): {dose_max_mm}')
+            print(
+                f'D_max RT/CT coord abs(\u0394): {abs(ct_max_mm - dose_max_mm)}')
+            print(f'1/2 CT voxel dimensions : {ct_voxel_size / 2}')
+
+    return (volume_max, ct_max_mm, dose_max_mm, ct_voxel_size)
+
+
+def d_max_check(path, volume, printing):
+    """
+    Function
+    ----------
+    Calculates the necessary offset for aligning the Dose Max value
+
+    Parameters
+    ----------
+    patient_path : str
+        A path directing towards a patient database in the following format:
+            MRN/[CT,PET,RTSTRUCT,RTDOSE]/*.dcm
+    dose_volume : array
+        An array of the dose volume, as returned form the dose
+            reconstruction function
+    printing : bool (Default = True)
+        Prints a formatting of the results, along with the voxel size
+
+    Returns
+    ----------
+    offset : np.array
+        An array of the voxel offset in the [x, y, z] axis
+    """
+    _, ct_max, dose_max, v_size = d_max_coords(path, volume, printing)
+    offset = [0, 0, 0]
+
+    for i, diff in enumerate(ct_max - dose_max):
+        pad = 30  # Chosen because > 30 offset likely indicates greater issues
+        low, high, spacing = (-pad * v_size[i], pad * v_size[i], v_size[i])
+        offset_list = np.arange(low, high, spacing)
+        offset[i] = nearest(offset_list, diff) - pad
+
+    return np.array(offset, dtype=int)
+
+
+def find_series_slices(path, find_associated=False):
+    """
+    Function
+    ----------
+    Given the path to a slice of a volume, checks all files within the
+        directory for the remaining slices of a given volume using
+        the SeriesInstanceUID
+
+    Parameters
+    ----------
+    path : pathlib.Path
+        A path to a DICOM image volume slice
+    find_associated : bool (Default = False)
+        A boolean designating finding a volume associated to a
+            associated RTSTRUCT based on SeriesInstanceUID
+
+    Returns
+    ----------
+    volume_slice_list : list
+        A list of files belonging to an image volume
+    """
+    if not isinstance(path, Path):
+        path = Path(path)
+
+    modality_list = ['CT', 'MR', 'PET', 'CBCT', 'RTSTRUCT']
+
+    if path.is_dir():
+        dir_contents = list(path.iterdir())
+        if dir_contents[0].is_file():
+            if dir_contents[0].suffix == '.dcm':
+                seed_file = dir_contents[0]
+            else:
+                print(f"{dir_contents}")
+                raise ValueError("Please provide a path to a DICOM File")
+        elif dir_contents[0].is_dir():
+            for d in dir_contents:
+                if d.is_dir():
+                    if d.name in modality_list:
+                        print("Using first volume/RTSTRUCT found")
+                        d_sub = [f for f in d.iterdir() if f.suffix == '.dcm']
+                        seed_file = d_sub[0]
+                        print(f"{seed_file}")
+                        break
+                    else:
+                        print("Using first dicom found")
+                        d_globs = list(d.rglob('*.dcm'))
+                        seed_file = d_globs[0]
+                        print(f"{seed_file}")
+                        break
+                elif d.is_file() and d.suffix == ".dcm":
+                    seed_file = d
+                    print(f"{seed_file}")
+                    break
+        else:
+            print("Check path")
+            return None
+    elif path.is_file():
+        seed_file = path
+    else:
+        raise ValueError("Path is not a file or a directory")
+
+    if find_associated:
+        dcmheader = pydicom.dcmread(str(seed_file), stop_before_pixels=True)
+        if 'ReferencedFrameOfReferenceSequence' in dir(dcmheader):
+            target_frame_of_reference = dcmheader.ReferencedFrameOfReferenceSequence[
+                0].FrameOfReferenceUID
+            pt_folders = list(path.parent.parent.iterdir())
+            for p in pt_folders:
+                if p.name in modality_list:
+                    candidate_file = next(p.iterdir())
+                    target_header = pydicom.dcmread(
+                        str(candidate_file), stop_before_pixels=True)
+                    if 'FrameOfReferenceUID' in dir(target_header):
+                        if target_header.FrameOfReferenceUID == target_frame_of_reference:
+                            seed_file = candidate_file
+                            break
+                        else:
+                            pass
+        else:
+            print("No Frame of Reference Found")
+            return None
+
+    file_list = list(seed_file.parent.iterdir())
+
+    volume_slice_list = []
+    series = []
+    for f in file_list:
+        dcmheader = pydicom.dcmread(str(f), stop_before_pixels=True)
+        if not series:
+            series = dcmheader.SeriesInstanceUID
+            volume_slice_list.append(str(f))
+        elif dcmheader.SeriesInstanceUID == series:
+            volume_slice_list.append(str(f))
+
+    volume_slice_list.sort()
+
+    return volume_slice_list
+
+
+def show_volume(volume_array, rows=2, cols=6, figsize=None):
+    """
+    Function
+    ----------
+    Displays the given image volume
+
+    Paramters
+    ----------
+    volume_array : np.array
+        Numpy arrays designating image and contour volumes
+    rows, cols : int (Default: rows = 2, cols = 6)
+        Intergers representing the rows, columns of the displayed images
+    figsize : int (Default = None)
+        An integer representing the pyplot figure size
+    """
+    if not figsize:
+        figsize = (cols * 4, rows * 3)
+
+    plt.figure(figsize=figsize)
+    zmax = volume_array.shape[-1]
+    imgs_out = rows * cols
+    skip = zmax // imgs_out
+    extra = zmax % imgs_out
+    for i in range(imgs_out):
+        plt.subplot(rows, cols, i+1)
+        z = i * skip + int(extra/2 + skip/2)
+        plt.imshow(volume_array[:, :, z], cmap='gray', interpolation='none')
+        plt.axis('off')
+        plt.title(f"z={z}")
+    plt.show
+
+
+def show_contours(volume_arr, contour_arr, rows=2, cols=6,
+                  orientation='axial', aspect=3, figsize=None):
+    """
+    Function
+    ----------
+    Displays a volume with overlaid contours
+
+    Paramters
+    ----------
+    volume_arr, contour_arr : np.array
+        Numpy arrays designating image and contour volumes
+    rows, cols : int (Default: rows = 2, cols = 2)
+        Intergers representing the rows, columns of the displayed images
+    aspect : float (Default = 3)
+        An aspect ratio for image display
+    orientation : str (Default = 'axial')
+        Designates the axis from which to display the images. Options include
+            'axial', 'sagittal' and 'coronal'
+    figsize : int (Default = None)
+        An integer representing the pyplot figure size
+    """
+
+    # Currently hard coded the aspect ratio for MRI context. May need to adapt it to make it more adaptible.
+
+    if not figsize:
+        figsize = (cols * 4, rows * 3)
+
+    masked_contour_arr = np.ma.masked_equal(contour_arr, 0)
+
+    plt.figure(figsize=figsize)
+    xmax, ymax, zmax = volume_arr.shape
+    imgs_out = rows * cols
+
+    if orientation == "axial":
+        max_dim = zmax
+        dim = "z"
+    elif orientation == "coronal":
+        max_dim = xmax
+        dim = "x"
+    elif orientation == "sagittal":
+        max_dim = ymax
+        dim = "y"
+    else:
+        raise NameError('The orientation provided is not an option')
+
+    skip = max_dim // imgs_out
+    extra = max_dim % imgs_out
+    slice_index = list(range(int(skip/2 + extra/2),
+                             max_dim - int(skip/2 + extra/2), skip))
+
+    for i, s in enumerate(slice_index):
+        plt.subplot(rows, cols, i+1)
+
+        if orientation == "axial":
+            plt.imshow(volume_arr[:, :, s], cmap='gray', interpolation='none')
+            plt.imshow(masked_contour_arr[:, :, s],
+                       cmap='Pastel1', interpolation='none', alpha=0.5)
+        elif orientation == "coronal":
+            plt.imshow(np.rollaxis(volume_arr[s, :, ::-1], 1, 0), cmap='gray',
+                       interpolation='none', aspect=aspect)
+            plt.imshow(np.rollaxis(masked_contour_arr[s, :, ::-1], 1, 0),
+                       cmap='Pastel1', interpolation='none', alpha=0.5, aspect=aspect)
+        elif orientation == "sagittal":
+            plt.imshow(np.rollaxis(volume_arr[:, s, ::-1], 1, 0), cmap='gray',
+                       interpolation='none', aspect=aspect)
+            plt.imshow(np.rollaxis(masked_contour_arr[:, s, ::-1], 1, 0),
+                       cmap='Pastel1', interpolation='none', alpha=0.5, aspect=aspect)
+
+        plt.axis('off')
+        plt.title(f"{dim}={s}")
+
+    plt.show()
